QuanLyKyTucXa_API/requirements.txt
Flask
mysql-connector-python
redis
python-dotenv